api/requirements.txt
django==2.2

djangorestframework==3.8.1
djangorestframework-jwt 
psycopg2
uwsgi
django-cors-headers
django-paypal
Pillow
django-bootstrap3
celery
django-environ
boto3
redis
PyJWT==1.5.3